feat: add request Get user by id with his posts, profile, memberType
