+++ src/routes/graphql/index.ts
@@ -7,7 +7,7 @@ import {
 } from 'graphql';
 import { memberTypeQuery, memberTypesQuery, postQuery, postsQuery, profileQuery, profilesQuery, userQuery, usersQuery } from './queries/queries';
 import { createUserMutation } from './mutations';
-import { userProfilePostMemberByIdQuery, usersProfilesPostsMembersQuery, usersWithPostsProfileMemberTypeQuery } from './queries';
+import { userProfilePostMemberByIdQuery, usersProfilesPostsMembersQuery, usersWithPostsProfileMemberTypeQuery, userWithPostsProfileMemberTypeQuery } from './queries';
 
 
 const queries = new GraphQLObjectType({
@@ -28,6 +28,7 @@ const queries = new GraphQLObjectType({
     usersProfilesPostsMembers: usersProfilesPostsMembersQuery,
     userProfilePostMemberById: userProfilePostMemberByIdQuery,
     usersWithPostsProfileMemberType: usersWithPostsProfileMemberTypeQuery,
+    userWithPostsProfileMemberType: userWithPostsProfileMemberTypeQuery,
   }
 });
