+++ src/routes/graphql/queries/index.ts
@@ -1,5 +1,6 @@
 import userProfilePostMemberByIdQuery from "./userProfilePostMemberById";
 import usersProfilesPostsMembersQuery from "./usersProfilesPostsMembers";
-import usersWithPostsProfileMemberTypeQuery from "./userWithPostsProfileMemberType";
+import usersWithPostsProfileMemberTypeQuery from "./usersWithPostsProfileMemberType";
+import userWithPostsProfileMemberTypeQuery from "./userWithPostsProfileMemberType";
 
-export { userProfilePostMemberByIdQuery, usersProfilesPostsMembersQuery, usersWithPostsProfileMemberTypeQuery };
+export { userProfilePostMemberByIdQuery, usersProfilesPostsMembersQuery, userWithPostsProfileMemberTypeQuery, usersWithPostsProfileMemberTypeQuery };
